refactor(workflow): add prefix to jooq generated tables
