+++ core/shared/schema/music/repo/src/main/java/com/github/durex/music/repository/MusicRepository.java
@@ -1,6 +1,6 @@
 package com.github.durex.music.repository;
 
-import static com.github.durex.api.tables.Music.MUSIC;
+import static com.github.durex.api.tables.TMusic.MUSIC;
 
 import com.github.durex.music.api.Music;
 import com.github.durex.music.mapper.MusicMapper;
