+++ core/shared/schema/music/repo/src/main/java/com/github/durex/music/repository/PlayListRepository.java
@@ -1,6 +1,6 @@
 package com.github.durex.music.repository;
 
-import static com.github.durex.api.Tables.PLAYLIST;
+import static com.github.durex.api.tables.TPlaylist.PLAYLIST;
 
 import com.github.durex.music.api.PlayList;
 import com.github.durex.music.mapper.PlayListMapper;
